SIP-1350|minor renamings
