+++ sip-core/src/main/java/de/ikor/sip/foundation/core/framework/FrameworkAutoConfig.java
@@ -1,7 +1,7 @@
 package de.ikor.sip.foundation.core.framework;
 
 import de.ikor.sip.foundation.core.framework.routers.RouteStarter;
-import de.ikor.sip.foundation.core.framework.routers.RoutersAnnotationBasedFactory;
+
 import de.ikor.sip.foundation.core.framework.templates.FromCentralRouterTemplate;
 import de.ikor.sip.foundation.core.framework.templates.FromDirectOutConnectorRouteTemplate;
 import de.ikor.sip.foundation.core.framework.templates.FromMiddleComponentRouteTemplate;
@@ -9,7 +9,6 @@ import org.springframework.context.annotation.Import;
 
 @Import({ // TODO load whole package
   RouteStarter.class,
-  RoutersAnnotationBasedFactory.class,
   FromMiddleComponentRouteTemplate.Template.class,
   FromDirectOutConnectorRouteTemplate.Template.class,
   FromCentralRouterTemplate.Template.class
