--- sip-core/src/main/java/de/ikor/sip/foundation/core/framework/definitions/SupremeRouteDefinition.java
@@ -1,8 +1,0 @@
-package de.ikor.sip.foundation.core.framework.definitions;
-
-import org.apache.camel.model.RouteDefinition;
-
-public class SupremeRouteDefinition {
-    private String endpointUri;
-    private RouteDefinition definition;
-}
