--- sip-core/src/main/java/de/ikor/sip/foundation/core/framework/routers/RoutersAnnotationBasedFactory.java
@@ -1,19 +1,0 @@
-package de.ikor.sip.foundation.core.framework.routers;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Service
-@RequiredArgsConstructor
-public class RoutersAnnotationBasedFactory {
-  private final List<CentralRouterDefinition> centralRouters;
-
-  public List<CentralRouterDefinition> getRouters() {
-    return centralRouters.stream()
-        .filter(router -> router.getClass().isAnnotationPresent(CentralRouterDomainModel.class))
-        .collect(Collectors.toList());
-  }
-}
